requirements.txt
alembic==1.8.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
Flask==2.1.2
Flask-Bcrypt==1.0.1
flask-form==1.0.0
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
Pillow==9.1.1
pycparser==2.21
SQLAlchemy==1.4.37
Werkzeug==2.1.2
WTForms==3.0.1
